Remove old packages from apps feature
K3sup used to have an app installation feature, which moved
into arkade a long time ago. These packages are no longer
used or required and can be removed.
