+++ cmd/install_test.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/alexellis/k3sup/pkg/helm"
 	"golang.org/x/crypto/ssh"
 )
 
@@ -129,15 +128,6 @@ func Test_RewriteKubeconfig(t *testing.T) {
 
 	if len(match) != len(expectedContextsToReplace) {
 		t.Fatalf("unexpected error, got: %q, want: %q.", len(match), len(expectedContextsToReplace))
-	}
-}
-
-func Test_getHelmURL(t *testing.T) {
-	got := helm.GetHelmURL("amd64", "darwin", "v2.14.3")
-	want := "https://get.helm.sh/helm-v2.14.3-darwin-amd64.tar.gz"
-
-	if want != got {
-		t.Errorf("incorrect Helm URL - want: %s, but got: %s", want, got)
 	}
 }
